+++ Sources/StreamChat/Query/ChannelListQuery.swift
@@ -4,13 +4,82 @@
 
 import Foundation
 
+/// A query is used for querying specific channels from backend.
+/// You can specify filter, sorting, pagination, limit for fetched messages in channel and other options.
+public struct ChannelListQuery: Encodable, LocalConvertibleSortingQuery {
+    private enum CodingKeys: String, CodingKey {
+        case filter = "filter_conditions"
+        case sort
+        case user = "user_details"
+        case state
+        case watch
+        case presence
+        case pagination
+        case messagesLimit = "message_limit"
+        case membersLimit = "member_limit"
+    }
+
+    /// A filter for the query (see `Filter`).
+    public let filter: Filter<ChannelListFilterScope>
+    /// A sorting for the query (see `Sorting`).
+    public let sort: [Sorting<ChannelListSortingKey>]
+    /// A pagination.
+    public var pagination: Pagination
+    /// A number of messages inside each channel.
+    public let messagesLimit: Int
+    /// Number of members inside each channel.
+    public let membersLimit: Int
+    /// Query options.
+    public var options: QueryOptions = [.watch]
+
+    /// Init a channels query.
+    /// - Parameters:
+    ///   - filter: a channels filter.
+    ///   - sort: a sorting list for channels.
+    ///   - pageSize: a page size for pagination.
+    ///   - messagesLimit: a number of messages for the channel to be retrieved.
+    public init(
+        filter: Filter<ChannelListFilterScope>,
+        sort: [Sorting<ChannelListSortingKey>] = [],
+        pageSize: Int = .channelsPageSize,
+        messagesLimit: Int = .messagesPageSize,
+        membersLimit: Int = .channelMembersPageSize
+    ) {
+        self.filter = filter
+        self.sort = sort
+        pagination = Pagination(pageSize: pageSize)
+        self.messagesLimit = messagesLimit
+        self.membersLimit = membersLimit
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(filter, forKey: .filter)
+
+        if !sort.isEmpty {
+            try container.encode(sort, forKey: .sort)
+        }
+
+        try container.encode(messagesLimit, forKey: .messagesLimit)
+        try container.encode(membersLimit, forKey: .membersLimit)
+        try options.encode(to: encoder)
+        try pagination.encode(to: encoder)
+    }
+}
+
+extension ChannelListQuery: CustomDebugStringConvertible {
+    public var debugDescription: String {
+        "Filter: \(filter) | Sort: \(sort)"
+    }
+}
+
 /// A namespace for the `FilterKey`s suitable to be used for `ChannelListQuery`. This scope is not aware of any extra data types.
 public protocol AnyChannelListFilterScope {}
 
 /// An extra-data-specific namespace for the `FilterKey`s suitable to be used for `ChannelListQuery`.
 public struct ChannelListFilterScope: FilterScope, AnyChannelListFilterScope {}
 
-public extension Filter where Scope: AnyChannelListFilterScope {
+public extension Filter where Scope == ChannelListFilterScope {
     /// Filter to match channels containing members with specified user ids.
     static func containMembers(userIds: [UserId]) -> Filter<Scope> {
         .in(.members, values: userIds)
@@ -32,7 +101,7 @@ public extension Filter where Scope: AnyChannelListFilterScope {
     }
 }
 
-extension Filter where Scope: AnyChannelListFilterScope {
+extension Filter where Scope == ChannelListFilterScope {
     /// Computed var helping us determine the value of `hidden` filter.
     var hiddenFilterValue: Bool? {
         if `operator`.isGroupOperator {
@@ -54,7 +123,7 @@ public enum InviteFilterValue: String, FilterValue {
 }
 
 /// Filter keys for channel list.
-public extension FilterKey where Scope: AnyChannelListFilterScope {
+public extension FilterKey where Scope == ChannelListFilterScope {
     /// A filter key for matching the `cid` value.
     /// Supported operators: `in`, `equal`
     static var cid: FilterKey<Scope, ChannelId> { .init(rawValue: "cid", keyPathString: #keyPath(ChannelDTO.cid), valueMapper: { $0.rawValue }) }
@@ -211,7 +280,7 @@ public extension FilterKey where Scope: AnyChannelListFilterScope {
 
 /// Internal filter queries for the channel list.
 /// These ones are helpers that should be used by an higher-level filter.
-internal extension FilterKey where Scope: AnyChannelListFilterScope {
+internal extension FilterKey where Scope == ChannelListFilterScope {
     /// Filter for fetching only the unread channels.
     /// Supported operators: `equal`, and only `true` is supported.
     static var hasUnread: FilterKey<Scope, Bool> {
@@ -228,74 +297,5 @@ internal extension FilterKey where Scope: AnyChannelListFilterScope {
                 }
             }
         )
-    }
-}
-
-/// A query is used for querying specific channels from backend.
-/// You can specify filter, sorting, pagination, limit for fetched messages in channel and other options.
-public struct ChannelListQuery: Encodable {
-    private enum CodingKeys: String, CodingKey {
-        case filter = "filter_conditions"
-        case sort
-        case user = "user_details"
-        case state
-        case watch
-        case presence
-        case pagination
-        case messagesLimit = "message_limit"
-        case membersLimit = "member_limit"
-    }
-
-    /// A filter for the query (see `Filter`).
-    public let filter: Filter<ChannelListFilterScope>
-    /// A sorting for the query (see `Sorting`).
-    public let sort: [Sorting<ChannelListSortingKey>]
-    /// A pagination.
-    public var pagination: Pagination
-    /// A number of messages inside each channel.
-    public let messagesLimit: Int
-    /// Number of members inside each channel.
-    public let membersLimit: Int
-    /// Query options.
-    public var options: QueryOptions = [.watch]
-
-    /// Init a channels query.
-    /// - Parameters:
-    ///   - filter: a channels filter.
-    ///   - sort: a sorting list for channels.
-    ///   - pageSize: a page size for pagination.
-    ///   - messagesLimit: a number of messages for the channel to be retrieved.
-    public init(
-        filter: Filter<ChannelListFilterScope>,
-        sort: [Sorting<ChannelListSortingKey>] = [],
-        pageSize: Int = .channelsPageSize,
-        messagesLimit: Int = .messagesPageSize,
-        membersLimit: Int = .channelMembersPageSize
-    ) {
-        self.filter = filter
-        self.sort = sort
-        pagination = Pagination(pageSize: pageSize)
-        self.messagesLimit = messagesLimit
-        self.membersLimit = membersLimit
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(filter, forKey: .filter)
-
-        if !sort.isEmpty {
-            try container.encode(sort, forKey: .sort)
-        }
-
-        try container.encode(messagesLimit, forKey: .messagesLimit)
-        try container.encode(membersLimit, forKey: .membersLimit)
-        try options.encode(to: encoder)
-        try pagination.encode(to: encoder)
-    }
-}
-
-extension ChannelListQuery: CustomDebugStringConvertible {
-    public var debugDescription: String {
-        "Filter: \(filter) | Sort: \(sort)"
     }
 }